auth_service/app/requirements.txt
Flask==2.2.3
Werkzeug==2.2.3
docker
pymongo==4.3.3
PyJWT==2.6.0
bcrypt
